test(scene-editor): split create-scene specs by workflow, tags, and submission
